update button in title bar when update exe is pending
